fix:float to int.

+++ src/Xo.AzDO.Engine/Models/Internal/Cmds/UpdateWiCmd.cs
@@ -3,5 +3,5 @@ namespace Xo.AzDO.Engine.Models;
 public class UpdateWiCmd : BaseWiCmd, IProcessorCmd
 {
     public int id { get; init; }
-    public float complete { get; init; }
+    public int complete { get; init; }
 }
